fix: remove networkUpload2
