+++ src/App.js
@@ -9,11 +9,11 @@ import OwnNetworksPage from './page/OwnNetworksPage/Index';
 import MyNetworksPage from './page/MyNetworksPage/Index';
 import MyInfoPage from './page/MyInfoPage/Index';
 import SignupPage from './page/SignupPage/Index';
-import NetworkUploadPage from './page/NetworkUploadPage2/Index';
+import NetworkUploadPage from './page/NetworkUploadPage/Index';
 import NetworkUpdatePage from './page/NetworkUpdatePage/Index';
 import MyInfoRevisePage from './page/MyInfoRevisePage/Index';
 import NetworkDetailPage from './page/NetworkDetailPage/Index';
-import { WeightUploadForm2 } from './components/NetworkUpload2';
+import { WeightUploadForm } from './components/NetworkUpload';
 import MySubscribePage from './page/MySubscribePage/Index';
 import Footer from './components/Footer/Index';
 import IntroductionPage from './page/IntroductionPage/Index';
@@ -46,7 +46,7 @@ function App() {
 					<Route path="/NetworkUpload" element={<NetworkUploadPage />} />
 				</Route>
 				<Route path="/WeightUpload" element={<AuthRoute />}>
-					<Route path="/WeightUpload" element={<WeightUploadForm2 />} />
+					<Route path="/WeightUpload" element={<WeightUploadForm />} />
 				</Route>
 				<Route path="/NetworkDetail/:postId" element={<NetworkDetailPage />} />
 				<Route path="/NetworkUpdate/:postId" element={<AuthRoute />}>
