--- src/components/NetworkUpload2/index.js
@@ -1,3 +1,0 @@
-export { default as InferenceForm2 } from './inferenceForm';
-export { default as NetworkDetailForm2 } from './networkDetailForm2';
-export { default as WeightUploadForm2 } from './weightUploadForm2';
